requirements.txt
antlr4-python3-runtime==4.8
backcall==0.2.0
black==21.8b0
block.bootstrap.pytorch==0.1.6
bootstrap.pytorch==0.0.13
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
easydict==1.9
einops==0.3.2
filelock==3.0.12
flake8==3.9.2
huggingface-hub==0.0.16
hydra-core==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mccabe==0.6.1
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.19.5
omegaconf==2.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
plotly==5.3.1
pretrainedmodels==0.7.4
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
skipthoughts==0.0.1
tabulate==0.8.9
tenacity==8.0.1
tensorboardX==2.4
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
traitlets==4.3.3
transformers==4.10.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
xxhash==2.0.2
zipp==3.5.0
